+++ src/main/java/eu/xenit/gradle/docker/compose/DockerComposePlugin.java
@@ -2,6 +2,7 @@ package eu.xenit.gradle.docker.compose;
 
 import com.avast.gradle.dockercompose.ComposeSettings;
 import eu.xenit.gradle.docker.config.DockerConfigPlugin;
+import eu.xenit.gradle.docker.internal.GradleVersionRequirement;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 
@@ -13,6 +14,7 @@ public class DockerComposePlugin implements Plugin<Project> {
 
     @Override
     public void apply(Project project) {
+        GradleVersionRequirement.atLeast("7.0", String.format("use the %s plugin", PLUGIN_ID));
         dockerComposeConvention = project.getExtensions()
                 .create("dockerCompose", DockerComposeExtensionOverride.class, project);
         project.getPluginManager().apply(DockerConfigPlugin.class);